+++ nzshm_model/__init__.py
@@ -1,13 +1,13 @@
-from . import nshm_1_0_0, nshm_1_0_4
+from . import nshm_v1_0_0, nshm_v1_0_4
 
 # Python package version is different than the NSHM MODEL version !!
-__version__ = '0.2.0'
+__version__ = '0.3.0'
 
-CURRENT_VERSION = "NSHM_1.0.0"
+CURRENT_VERSION = "NSHM_v1.0.0"
 
 versions = {
-    "NSHM_1.0.0": nshm_1_0_0,
-    "NSHM_1.0.4": nshm_1_0_4,
+    "NSHM_v1.0.0": nshm_v1_0_0,
+    "NSHM_v1.0.4": nshm_v1_0_4,
 }